Remove dependency on @octokit/webhooks
This dependency was added, mainly, because of types. The problem is that
when Probot updates its internal @octokit/webhooks dependency, there is
a version mismatch and we need to also update ours, which makes it
harder to update probot. Thus, removing it and falling back to
typescript implicit typings and creating interfaces for typing functions
to make linters happy.